Factor out AreAnalysisScopeMatching
