+++ analyzers/src/SonarAnalyzer.Common/Helpers/CompilationExtensions.cs
@@ -18,6 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using System.Collections.Generic;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 
@@ -34,5 +35,19 @@ namespace SonarAnalyzer.Helpers
             // There's no direct way of checking compilation target framework yet (09/2020).
             // See https://github.com/dotnet/roslyn/issues/3798
             compilation.ObjectType.ContainingAssembly.Name == "mscorlib";
+
+        internal static bool AreAnalysisScopeMatching(this Compilation compilation, IEnumerable<DiagnosticDescriptor> diagnostics)
+        {
+            if (compilation == null)
+            {
+                return true; // We don't know whether this is a Main or Test source so let's run the rule
+            }
+
+            var matchingScopeTag = compilation.IsTest()
+                ? DiagnosticDescriptorBuilder.TestSourceScopeTag
+                : DiagnosticDescriptorBuilder.MainSourceScopeTag;
+
+            return diagnostics.Any(d => d.CustomTags.Contains(matchingScopeTag));
+        }
     }
 }
